--- src/main/java/com/example/techit7/comment/CommentServiceImpl.java
@@ -1,30 +1,0 @@
-package com.example.techit7.comment;
-
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-
-@Service
-public class CommentServiceImpl implements CommentService {
-
-
-    @Override
-    public List<Comment> getComments() {
-        return null;
-    }
-
-    @Override
-    public void postComment() {
-
-    }
-
-    @Override
-    public void updateComment(Long id) {
-
-    }
-
-    @Override
-    public void deleteComment(Long id) {
-
-    }
-}
