+++ src/main/java/com/example/techit7/comment/dto/CommentRequestDto.java
@@ -1,0 +1,4 @@
+package com.example.techit7.comment.dto;
+
+public class CommentRequestDto {
+}
