--- src/main/java/com/example/techit7/domain/Article.java
@@ -1,34 +1,0 @@
-package com.example.techit7.domain;
-
-import jakarta.persistence.*;
-import java.util.Date;
-import java.util.List;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.Setter;
-
-@Entity
-@Getter @Setter
-@RequiredArgsConstructor
-public class Article {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "article_id")
-    private long id;
-
-    @ManyToOne(targetEntity = User.class, fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    private User author;
-
-    private String title;
-    private String content;
-    private Date regDate;
-    private Date modDate;
-    private long viewCount;
-    private long likeCount;
-    private String category;
-
-    @OneToMany(mappedBy = "article", fetch = FetchType.LAZY,cascade = CascadeType.REMOVE)
-    private List<Comment> commentList;
-
-}
