--- src/main/java/com/example/techit7/domain/ArticleVote.java
@@ -1,25 +1,0 @@
-package com.example.techit7.domain;
-
-import com.example.techit7.article.Article;
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.Setter;
-
-@Entity
-@Getter @Setter
-@RequiredArgsConstructor
-public class ArticleVote {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "articlevote_id")
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    private User user; //author?
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "article_id")
-    private Article article;
-}
