--- src/main/java/com/example/techit7/domain/CommentVote.java
@@ -1,24 +1,0 @@
-package com.example.techit7.domain;
-
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import jakarta.persistence.*;
-
-@Entity
-@Getter
-@RequiredArgsConstructor
-public class CommentVote {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "like_comment_id")
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id")
-    private User user;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "comment_id")
-    private Comment comment;
-}
