--- src/main/java/com/example/techit7/domain/User.java
@@ -1,30 +1,0 @@
-package com.example.techit7.domain;
-
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.Setter;
-import jakarta.persistence.*;
-
-@Entity
-@Getter @Setter
-@RequiredArgsConstructor
-public class User {
-  
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column
-    private String loginId;
-
-    @Column
-    private String password;
-
-    @Column
-    private String nickname;
-
-    @Column
-    private String email;
-
-}
